pyproject.toml
[tool.poetry]
name = "pip-search"
version = "0.1.0"
description = "Script to search pip packages via http"
authors = ["quinn-n <quinnneufeld@protonmail.com>"]
license = "WTFPL"
readme = "README.md"
packages = [{include = "pip_search"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
click = "^8.1.3"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
attrs = "^22.1.0"
colorama = "^0.4.5"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
pytest = "^7.1.3"
mypy = "^0.971"
ipython = "^8.5.0"
hypothesis = "^6.54.6"
pyinstaller = "^5.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
